pyproject.toml
[tool.poetry]
name = "farmrpg-etl2"
version = "0.1.0"
description = ""
authors = ["Noah Kantrowitz <noah@coderanger.net>"]
readme = "README.md"
packages = [
    { include = "farmrpg_etl", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.1.5"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
attrs = "^23.1.0"
structlog = "^23.1.0"
strawberry-django-plus = "^2.3.2"
django-debug-toolbar = "^4.0.0"
strawberry-graphql-django = "^0.9.4"
dj-database-url = "^2.0.0"
psycopg2 = "^2.9.5"
httpx = "^0.24.0"
djangorestframework = "^3.14.0"
django-pghistory = "^2.5.1"
lxml = "^4.9.2"
html5lib = "^1.1"
cssselect = "^1.2.0"
lxml-stubs = "^0.4.0"
toposort = "^1.10"
whitenoise = "^6.4.0"
beautifulsoup4 = "^4.11.2"
google-cloud-firestore = "^2.10.0"
firebase-admin = "^6.1.0"
google-cloud-iam = "^2.12.0"
async-lru = "^2.0.2"
numpy = "^1.24.3"
sentry-sdk = "^1.21.1"
python-dateutil = "^2.8.2"
croniter = "^1.4.1"

[tool.poetry.group.dev.dependencies]
respx = "^0.20.1"
pytest = "^7.2.2"
pytest-asyncio = "^0.21.0"
pytest-django = "^4.5.2"
factory-boy = "^3.2.1"
ipython = "^8.13.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "farmrpg_etl.settings"
python_files = "tests.py test_*.py *_tests.py"
